perp before fork

+++ functionlist.h
@@ -3,7 +3,7 @@
 
 #define pi 3.14159
 
-unsigned long time1, time2;
+float time1, time2;
 bool motorsEnabled;
